+++ Backend/src/services/productService.js
@@ -1,6 +1,8 @@
 const createError = require("http-errors");
 
 const { Product } = require("../models/product.model");
+const cloudinary = require("../config/cloudinary");
+const { imagePublicUrlWithoutExtention } = require("../helper/cloudinary");
 
 const handleCreateProductService = async (productData) => {
   try {
@@ -10,16 +12,22 @@ const handleCreateProductService = async (productData) => {
       throw createError(400, "Product is already exist.");
     }
 
-    const product = await Product.create(productData);
+    const ExistingProduct = productData;
+    const response = await cloudinary.uploader.upload(ExistingProduct.image, {
+      folder: "e-commerce-mern/products"
+    });
+    ExistingProduct.image = response.secure_url;
+
+    const product = await Product.create(ExistingProduct);
     return product;
   } catch (error) {
     throw error;
   }
 };
 
-const handleReadAllProductService = async (page, limit) => {
+const handleReadAllProductService = async (page, limit, filter) => {
   try {
-    const products = await Product.find({})
+    const products = await Product.find(filter)
       .select("-image")
       .populate("category")
       .limit(limit)
@@ -30,7 +38,7 @@ const handleReadAllProductService = async (page, limit) => {
       throw createError(400, "Products Not found.");
     }
 
-    const count = await Product.find({}).countDocuments();
+    const count = await Product.find(filter).countDocuments();
     if (!count) {
       throw createError(400, "Products Not found.");
     }
@@ -54,6 +62,19 @@ const handleReadSingleProductService = async (slug) => {
 
 const handleDeleteSingleProductService = async (slug) => {
   try {
+    const productData = await Product.findOne({ slug });
+    if (!productData) {
+      throw createError(400, "Product is not found with this Slug");
+    }
+    const productImageWithoutExtention = imagePublicUrlWithoutExtention(
+      productData.image
+    );
+    const { result } = await cloudinary.uploader.destroy(
+      `e-commerce-mern/products/${productImageWithoutExtention}`
+    );
+    if (result !== "ok") {
+      throw createError(400, "Product image is not deleted. Please try again!");
+    }
     const product = await Product.findOneAndDelete({ slug });
     if (!product) {
       throw createError(400, "Product is Not found with this slug.");
